init test class and wiremock server

+++ src/main/java/dev/bazarski/clashqualifiers/records/Account.java
@@ -3,5 +3,5 @@ package dev.bazarski.clashqualifiers.records;
 public record Account(String puuid,
                       String gameName,
                       String tagLine,
-                      Integer initialPoints
+                      Double initialPoints
 ) {}
